Refactor for xpdo3/modx3
