+++ core/components/articles/lexicon/en/default.inc.php
@@ -56,7 +56,7 @@ $_lang['articles.author'] = 'Author';
 $_lang['articles.categories'] = 'Categories';
 $_lang['articles.comments'] = 'Comments';
 $_lang['articles.comments.intro_msg'] = 'Here you can review, approve and reject comments for all posts for your Articles.';
-$_lang['articles.container'] = 'Articles';
+$_lang['articles.container'] = 'Articles Container';
 $_lang['articles.container_alias'] = 'Container Alias';
 $_lang['articles.container_alias_desc'] = 'The web friendly URL that will be used to calculate the URL of this Articles Container.';
 $_lang['articles.container_create_here'] = 'Create Articles Here';
