--- core/components/articles/model/articles/metadata.mysql.php
@@ -1,9 +1,0 @@
-<?php
-
-$xpdo_meta_map = array (
-  'modResource' => 
-  array (
-    0 => 'ArticlesContainer',
-    1 => 'Article',
-  ),
-);
