--- core/components/articles/model/articles/mysql/article.map.inc.php
@@ -1,30 +1,0 @@
-<?php
-/**
- * @package articles
- */
-$xpdo_meta_map['Article']= array (
-  'package' => 'articles',
-  'version' => '1.1',
-  'extends' => 'modResource',
-  'tableMeta' => 
-  array (
-    'engine' => 'MyISAM',
-  ),
-  'fields' => 
-  array (
-  ),
-  'fieldMeta' => 
-  array (
-  ),
-  'aggregates' => 
-  array (
-    'Container' => 
-    array (
-      'class' => 'ArticlesContainer',
-      'local' => 'parent',
-      'foreign' => 'id',
-      'cardinality' => 'one',
-      'owner' => 'foreign',
-    ),
-  ),
-);
