--- core/components/articles/model/articles/mysql/articlescontainer.map.inc.php
@@ -1,30 +1,0 @@
-<?php
-/**
- * @package articles
- */
-$xpdo_meta_map['ArticlesContainer']= array (
-  'package' => 'articles',
-  'version' => '1.1',
-  'extends' => 'modResource',
-  'tableMeta' => 
-  array (
-    'engine' => 'MyISAM',
-  ),
-  'fields' => 
-  array (
-  ),
-  'fieldMeta' => 
-  array (
-  ),
-  'composites' => 
-  array (
-    'Articles' => 
-    array (
-      'class' => 'Article',
-      'local' => 'id',
-      'foreign' => 'parent',
-      'cardinality' => 'many',
-      'owner' => 'local',
-    ),
-  ),
-);
