+++ core/components/articles/src/Model/Article.php
@@ -1,33 +1,17 @@
 <?php
-/**
- * Articles
- *
- * Copyright 2011-12 by Shaun McCormick <shaun+articles@modx.com>
- *
- * Articles is free software; you can redistribute it and/or modify it under the
- * terms of the GNU General Public License as published by the Free Software
- * Foundation; either version 2 of the License, or (at your option) any later
- * version.
- *
- * Articles is distributed in the hope that it will be useful, but WITHOUT ANY
- * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
- * A PARTICULAR PURPOSE. See the GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along with
- * Articles; if not, write to the Free Software Foundation, Inc., 59 Temple
- * Place, Suite 330, Boston, MA 02111-1307 USA
- *
- * @package articles
- */
 
-/**
- * Detect if we are running MODX 2.x or 3.x and include the required files if we are on 2.x
- */
-if (!class_exists('\MODX\Revolution\modX')) {
-    require_once MODX_CORE_PATH . 'model/modx/modprocessor.class.php';
-    require_once MODX_CORE_PATH . 'model/modx/processors/resource/create.class.php';
-    require_once MODX_CORE_PATH . 'model/modx/processors/resource/update.class.php';
-}
+namespace Articles\Model;
+
+use Articles\Model\Notification\ArticlesNotification;
+use Articles\Model\Update\ArticlesPingomatic;
+use Articles\Model\Update\ArticlesUpdateService;
+use MODX\Revolution\modContentType;
+use MODX\Revolution\modResource;
+use MODX\Revolution\modResourceGroupResource;
+use MODX\Revolution\modTemplateVarResource;
+use MODX\Revolution\modX;
+use quipThread;
+use xPDO\xPDO;
 
 /**
  * @package articles
@@ -38,7 +22,7 @@ class Article extends modResource {
 
     function __construct(xPDO &$xpdo) {
         parent :: __construct($xpdo);
-        $this->set('class_key',Article::class);
+        $this->set('class_key',self::class);
         $this->set('show_in_tree',false);
         //$this->set('richtext',true);
         $this->set('searchable',true);
@@ -115,7 +99,7 @@ class Article extends modResource {
    &useCss=`'.$this->xpdo->getOption('commentsUseCss',$settings,1).'`
    &altRowCss=`'.$this->xpdo->getOption('commentsAltRowCss',$settings,'quip-comment-alt').'`
 
-   &zzrequireAuth=`'.$this->xpdo->getOption('commentsRequireAuth',$settings,0).'`
+   &requireAuth=`'.$this->xpdo->getOption('commentsRequireAuth',$settings,0).'`
 
    &nameField=`'.$this->xpdo->getOption('commentsNameField',$settings,'username').'`
    &showAnonymousName=`'.$this->xpdo->getOption('commentsShowAnonymousName',$settings,0).'`
@@ -202,7 +186,7 @@ class Article extends modResource {
         $settings = $this->getContainerSettings();
         if (empty($settings['notificationServices'])) return $success;
         $services = explode(',',$settings['notificationServices']);
-        $modelPath = $this->xpdo->getOption('articles.core_path',null,$this->xpdo->getOption('core_path').'components/articles/').'model/articles/';
+        $modelPath = $this->xpdo->getOption('articles.core_path',null,$this->xpdo->getOption('core_path').'components/articles/').'src/Model/';
 
         /* get context to prepare url mapping */
         $this->xpdo->getContext($this->get('context_key'));
@@ -241,13 +225,13 @@ class Article extends modResource {
 
     /**
      * Get the notification service
-     * @return ArticlesNotifyService
+     * @return ArticlesPingomatic
      */
     protected function getUpdateService() {
         $settings = $this->getContainerSettings();
-        $modelPath = $this->xpdo->getOption('articles.core_path',null,$this->xpdo->getOption('core_path').'components/articles/').'model/articles/';
-        $updateServiceClass = $this->getOption('updateServiceClass',$settings,'ArticlesPingomatic');
-        $updateServicePath = $this->getOption('updateServicePath',$settings,$modelPath.'update/articlespingomatic.class.php');
+        $modelPath = $this->xpdo->getOption('articles.core_path',null,$this->xpdo->getOption('core_path').'components/articles/').'src/Model/';
+        $updateServiceClass = $this->getOption('updateServiceClass',$settings,ArticlesPingomatic::class);
+        $updateServicePath = $this->getOption('updateServicePath',$settings,$modelPath.'Update/ArticlesPingomatic.php');
         $included = include_once $updateServicePath;
         $service = false;
         if ($included) {
@@ -444,304 +428,3 @@ class Article extends modResource {
     }
 }
 
-/**
- * Overrides the modResourceCreateProcessor to provide custom processor functionality for the Article type
- *
- * @package articles
- */
-class ArticleCreateProcessor extends modResourceCreateProcessor {
-    /** @var modResource $yearParent */
-    public $yearParent;
-    /** @var modResource $monthParent */
-    public $monthParent;
-    /** @var modResource $dayParent */
-    public $dayParent;
-    /** @var Article $object */
-    public $object;
-    /** @var boolean $isPublishing */
-    public $isPublishing = false;
-
-    public function beforeSet() {
-        $this->setProperty('searchable',true);
-        $this->setProperty('richtext',true);
-        $this->setProperty('isfolder',false);
-        $this->setProperty('cacheable',true);
-        $this->setProperty('clearCache',true);
-        $this->setProperty('class_key',Article::class);
-        return parent::beforeSet();
-    }
-
-    /**
-     * Override modResourceCreateProcessor::beforeSave to provide archiving
-     *
-     * {@inheritDoc}
-     * @return boolean
-     */
-    public function beforeSave() {
-        $beforeSave = parent::beforeSave();
-
-        if (!$this->parentResource) {
-            $this->parentResource = $this->object->getOne('Parent');
-        }
-
-        if ($this->object->get('published') || $this->object->get('pub_date')) {
-            if (!$this->setArchiveUri()) {
-                $this->modx->log(modX::LOG_LEVEL_ERROR,'Failed to set URI for new Article.');
-            }
-        }
-
-        /** @var ArticlesContainer $container */
-        $container = $this->modx->getObject(ArticlesContainer::class,$this->object->get('parent'));
-        if ($container) {
-            $settings = $container->getProperties('articles');
-            $this->object->setProperties($settings,'articles');
-            $this->object->set('richtext',!isset($settings['articlesRichtext']) || !empty($settings['articlesRichtext']));
-        }
-
-        $this->isPublishing = $this->object->isDirty('published') && $this->object->get('published');
-        return $beforeSave;
-    }
-
-    /**
-     * Set the friendly URL archive by forcing it into the URI.
-     * @return bool|string
-     */
-    public function setArchiveUri() {
-        if (!$this->parentResource) {
-            return false;
-        }
-        return $this->object->setArchiveUri();
-    }
-
-    public function afterSave() {
-        $afterSave = parent::afterSave();
-        $this->saveTemplateVariables();
-        if($this->object->get('clearCache')) $this->clearContainerCache();
-        if ($this->isPublishing) {
-            $this->object->notifyUpdateServices();
-            $this->object->sendNotifications();
-        }
-        return $afterSave;
-    }
-
-    /**
-     * Clears the container cache to ensure that the container listing is updated
-     * @return void
-     */
-    public function clearContainerCache() {
-        $this->modx->cacheManager->refresh([
-            'db' => [],
-            'auto_publish' => ['contexts' => [$this->object->get('context_key')]],
-            'context_settings' => ['contexts' => [$this->object->get('context_key')]],
-            'resource' => ['contexts' => [$this->object->get('context_key')]],
-        ]);
-    }
-
-    /**
-     * Extend the saveTemplateVariables method and provide handling for the 'tags' type to store in a hidden TV
-     * @return array|mixed
-     */
-    public function saveTemplateVariables() {
-        $tags = $this->getProperty('tags',null);
-        if ($tags !== null) {
-            /** @var modTemplateVar $tv */
-            $tv = $this->modx->getObject(modTemplateVar::class, [
-                'name' => 'articlestags',
-            ]);
-            if ($tv) {
-                $defaultValue = $tv->processBindings($tv->get('default_text'),$this->object->get('id'));
-                if (strcmp($tags,$defaultValue) != 0) {
-                    /* update the existing record */
-                    $tvc = $this->modx->getObject(modTemplateVarResource::class, [
-                        'tmplvarid' => $tv->get('id'),
-                        'contentid' => $this->object->get('id'),
-                    ]);
-                    if ($tvc == null) {
-                        /** @var modTemplateVarResource $tvc add a new record */
-                        $tvc = $this->modx->newObject(modTemplateVarResource::class);
-                        $tvc->set('tmplvarid',$tv->get('id'));
-                        $tvc->set('contentid',$this->object->get('id'));
-                    }
-                    $tvc->set('value',$tags);
-                    $tvc->save();
-
-                /* if equal to default value, erase TVR record */
-                } else {
-                    $tvc = $this->modx->getObject(modTemplateVarResource::class, [
-                        'tmplvarid' => $tv->get('id'),
-                        'contentid' => $this->object->get('id'),
-                    ]);
-                    if (!empty($tvc)) {
-                        $tvc->remove();
-                    }
-                }
-            }
-        }
-        return true;
-    }
-}
-
-/**
- * Overrides the modResourceUpdateProcessor to provide custom processor functionality for the Article type
- *
- * @package articles
- */
-class ArticleUpdateProcessor extends modResourceUpdateProcessor {
-    /** @var modResource $yearParent */
-    public $yearParent;
-    /** @var modResource $monthParent */
-    public $monthParent;
-    /** @var modResource $dayParent */
-    public $dayParent;
-    /** @var Article $object */
-    public $object;
-    /** @var boolean $isPublishing */
-    public $isPublishing = false;
-
-    public function beforeSet() {
-        $this->setProperty('clearCache',true);
-        return parent::beforeSet();
-    }
-
-    /**
-     * Override modResourceUpdateProcessor::beforeSave to provide archiving
-     *
-     * {@inheritDoc}
-     * @return boolean
-     */
-    public function beforeSave() {
-        $afterSave = parent::beforeSave();
-        $container = $this->modx->getObject(ArticlesContainer::class,$this->object->get('parent'));
-
-        if ($this->object->get('published') && ($this->object->isDirty('alias') || $this->object->isDirty('published'))) {
-            if (!$this->setArchiveUri()) {
-                $this->modx->log(modX::LOG_LEVEL_ERROR,'Failed to set date URI.');
-            }
-        } else if (($this->object->get('pub_date') && $this->object->isDirty('pub_date')) || $this->object->isDirty('pub_date')) {
-            if (!$this->setArchiveUri()) {
-                $this->modx->log(modX::LOG_LEVEL_ERROR,'Failed to set date URI pub_date.');
-            }
-        } else if(!$this->object->get('published') && !$this->object->get('pub_date')) { // we need to always do this because the url may have been set previously by pub_date
-	        /*$containerUri = $container->get('uri');
-	        if (empty($containerUri)) {
-	            $containerUri = $container->get('alias');
-	        }*/
-	        $uri = rtrim($this->object->get('alias'));
-	        $this->object->set('uri',$uri);
-	        $this->object->set('uri_override',true);
-        }
-
-        /** @var ArticlesContainer $container */
-        if ($container) {
-            $this->object->setProperties($container->getProperties('articles'),'articles');
-        }
-
-        $this->isPublishing = $this->object->isDirty('published') && $this->object->get('published');
-        return $afterSave;
-    }
-
-    /**
-     * Extend the saveTemplateVariables method and provide handling for the 'tags' type to store in a hidden TV
-     * @return array|mixed
-     */
-    public function saveTemplateVariables() {
-        $saved = parent::saveTemplateVariables();
-        $tags = $this->getProperty('tags',null);
-        if ($tags !== null) {
-            /** @var modTemplateVar $tv */
-            $tv = $this->modx->getObject(modTemplateVar::class, [
-                'name' => 'articlestags',
-            ]);
-            if ($tv) {
-                $defaultValue = $tv->processBindings($tv->get('default_text'),$this->object->get('id'));
-                if (strcmp($tags,$defaultValue) != 0) {
-                    /* update the existing record */
-                    $tvc = $this->modx->getObject(modTemplateVarResource::class, [
-                        'tmplvarid' => $tv->get('id'),
-                        'contentid' => $this->object->get('id'),
-                    ]);
-                    if ($tvc == null) {
-                        /** @var modTemplateVarResource $tvc add a new record */
-                        $tvc = $this->modx->newObject(modTemplateVarResource::class);
-                        $tvc->set('tmplvarid',$tv->get('id'));
-                        $tvc->set('contentid',$this->object->get('id'));
-                    }
-                    $tvc->set('value',$tags);
-                    $tvc->save();
-
-                /* if equal to default value, erase TVR record */
-                } else {
-                    $tvc = $this->modx->getObject(modTemplateVarResource::class, [
-                        'tmplvarid' => $tv->get('id'),
-                        'contentid' => $this->object->get('id'),
-                    ]);
-                    if (!empty($tvc)) {
-                        $tvc->remove();
-                    }
-                }
-            }
-        }
-        return $saved;
-    }
-
-    /**
-     * Set the friendly URL archive by forcing it into the URI.
-     * @return bool|string
-     */
-    public function setArchiveUri() {
-        if (!$this->parentResource) {
-            $this->parentResource = $this->object->getOne('Parent');
-            if (!$this->parentResource) {
-                return false;
-            }
-        }
-
-        return $this->object->setArchiveUri();
-    }
-
-    public function afterSave() {
-        $afterSave = parent::afterSave();
-        if ($this->isPublishing) {
-            $this->object->notifyUpdateServices();
-            $this->object->sendNotifications();
-        }
-        if($this->object->get('clearCache')) $this->clearContainerCache();
-        return $afterSave;
-    }
-
-    /**
-     * Clears the container cache to ensure that the container listing is updated
-     * @return void
-     */
-    public function clearContainerCache() {
-        $this->modx->cacheManager->refresh([
-            'db' => [],
-            'auto_publish' => ['contexts' => [$this->object->get('context_key')]],
-            'context_settings' => ['contexts' => [$this->object->get('context_key')]],
-            'resource' => ['contexts' => [$this->object->get('context_key')]],
-        ]);
-    }
-
-    /**
-     * Override cleanup to send only back needed params
-     * @return array|string
-     */
-    public function cleanup() {
-        $this->object->removeLock();
-        $this->clearCache();
-
-        $returnArray = $this->object->get(array_diff(array_keys($this->object->_fields), ['content','ta','introtext','description','link_attributes','pagetitle','longtitle','menutitle','articles_container_settings','properties']));
-        foreach ($returnArray as $k => $v) {
-            if (strpos($k,'tv') === 0) {
-                unset($returnArray[$k]);
-            }
-            if (strpos($k,'setting_') === 0) {
-                unset($returnArray[$k]);
-            }
-        }
-        $returnArray['class_key'] = $this->object->get('class_key');
-        $this->workingContext->prepare(true);
-        $returnArray['preview_url'] = $this->modx->makeUrl($this->object->get('id'), $this->object->get('context_key'), '', 'full');
-        return $this->success('',$returnArray);
-    }
-}
