+++ core/components/articles/src/Model/Notification/ArticlesNotification.php
@@ -19,8 +19,39 @@
  *
  * @package articles
  */
+
+namespace Articles\Model\Notification;
+
+use Articles\Model\Article;
+use MODX\Revolution\modX;
+
 /**
+ * Base abstract class for notification senders
+ *
  * @package articles
+ * @subpackage notifications
  */
-require_once (strtr(realpath(dirname(dirname(__FILE__))), '\\', '/') . '/article.class.php');
-class Article_mysql extends Article {}
+abstract class ArticlesNotification {
+    /** @var modX $xpdo */
+    public $modx;
+    /** @var Article $article */
+    public $article;
+    /** @var array $config */
+    public $config = [];
+
+    function __construct(Article $article,array $config = []) {
+        $this->article =& $article;
+        $this->modx =& $article->xpdo;
+        $this->config = array_merge([
+
+        ],$config);
+    }
+
+    /**
+     * @abstract
+     * @param string $title The title of the Article
+     * @param string $url The full URL of the Article
+     * @return boolean
+     */
+    abstract public function send($title,$url);
+}
