+++ core/components/articles/src/Model/mysql/ArticlesContainer.php
@@ -1,0 +1,53 @@
+<?php
+/**
+ * Articles
+ *
+ * Copyright 2011-12 by Shaun McCormick <shaun+articles@modx.com>
+ *
+ * Articles is free software; you can redistribute it and/or modify it under the
+ * terms of the GNU General Public License as published by the Free Software
+ * Foundation; either version 2 of the License, or (at your option) any later
+ * version.
+ *
+ * Articles is distributed in the hope that it will be useful, but WITHOUT ANY
+ * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+ * A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with
+ * Articles; if not, write to the Free Software Foundation, Inc., 59 Temple
+ * Place, Suite 330, Boston, MA 02111-1307 USA
+ *
+ * @package articles
+ */
+
+namespace Articles\Model\mysql;
+
+class ArticlesContainer extends \Articles\Model\ArticlesContainer
+{
+    public static $metaMap = array (
+        'package' => 'Articles\\Model\\',
+        'version' => '3.0',
+        'extends' => 'MODX\\Revolution\\modResource',
+        'tableMeta' =>
+            array (
+                'engine' => 'InnoDB',
+            ),
+        'fields' =>
+            array (
+            ),
+        'fieldMeta' =>
+            array (
+            ),
+        'composites' =>
+            array (
+                'Article' =>
+                    array (
+                        'class' => 'Articles\\Model\\Article',
+                        'local' => 'id',
+                        'foreign' => 'parent',
+                        'cardinality' => 'many',
+                        'owner' => 'local',
+                    ),
+            ),
+    );
+}
